Upgrade electron-builder and pin update-notifier
To address a CVE in dot-prop, we need to upgrade the package to >=
5.1.1. Unfortunately, it's a deeply nested dependency of
electron-builder. To get to 5.1.1, we would need to update to the latest
version of electron-builder (22.x.x). When attempting this update, `yarn
install` broken.

So, I just updated `electron-builder` to 21.x.x and pinned the
dependency that causes `dot-prop` to resolve to an older version of the
package. We had to pin somewhere; this seemed like minimal risk.

Also updates `webpack-dev-server` to address high level security
vulnerability.

[#174105633]